Fix circular imports
- Refactor config.go and io_utils.go into separate packages to resolve
circular imports within internal/crypto/openpgp
- Update dependencies with dep
