+++ internal/utils/io.go
@@ -1,9 +1,10 @@
-package internal
+package utils
 
 import (
-	"github.com/wal-g/wal-g/utility"
 	"io"
 	"os"
+
+	"github.com/wal-g/wal-g/utility"
 )
 
 type ReadSeekCloser interface {
